Fix some bugs

+++ H2CO3CustomKeyBoard/src/main/java/com/cainiaohh/module/h2co3customkeyboard/gamecontroller/codes/Translation.java
@@ -7,6 +7,6 @@ public class Translation {
     }
 
     public int trans(String str) {
-        return ((Integer) this.xKeyMap.translate(str)).intValue();
+        return (Integer) this.xKeyMap.translate(str);
     }
 }
